Fix typedefs of fmi3[U]Int{32,64}

+++ headers/fmi3TypesPlatform.h
@@ -66,10 +66,10 @@ typedef   int8_t        fmi3Int8;     /* 8-bit signed integer */
 typedef  uint8_t        fmi3UInt8;    /* 8-bit unsigned integer */
 typedef  int16_t        fmi3Int16;    /* 16-bit signed integer */
 typedef uint16_t        fmi3UInt16;   /* 16-bit unsigned integer */
-typedef  int16_t        fmi3Int32;    /* 32-bit signed integer */
-typedef uint16_t        fmi3UInt32;   /* 32-bit unsigned integer */
-typedef  int16_t        fmi3Int64;    /* 64-bit signed integer */
-typedef uint16_t        fmi3UInt64;   /* 64-bit unsigned integer */
+typedef  int32_t        fmi3Int32;    /* 32-bit signed integer */
+typedef uint32_t        fmi3UInt32;   /* 32-bit unsigned integer */
+typedef  int64_t        fmi3Int64;    /* 64-bit signed integer */
+typedef uint64_t        fmi3UInt64;   /* 64-bit unsigned integer */
 typedef int             fmi3Boolean;  /* Data type to be used with fmi3True and fmi3False */
 typedef char            fmi3Char;     /* Data type for one character */
 typedef const fmi3Char* fmi3String;   /* Data type for character strings
